+++ src/black/nodes.py
@@ -531,7 +531,7 @@ def is_arith_like(node: LN) -> bool:
     }
 
 
-def is_docstring(leaf: Leaf) -> bool:
+def is_docstring(leaf: Leaf, mode: Mode) -> bool:
     if leaf.type != token.STRING:
         return False
 
@@ -539,7 +539,14 @@ def is_docstring(leaf: Leaf) -> bool:
     if set(prefix).intersection("bBfF"):
         return False
 
-    if not leaf.parent.prev_sibling:
+    if (
+        Preview.format_module_docstring in mode
+        and leaf.parent
+        and leaf.parent.type == syms.simple_stmt
+        and not leaf.parent.prev_sibling
+        and leaf.parent.parent
+        and leaf.parent.parent.type == syms.file_input
+    ):
         return True
 
     if prev_siblings_are(